refactor: コードの整理
